refac: retrieval app structuring
