+++ backend/open_webui/apps/webui/routers/knowledge.py
@@ -4,6 +4,8 @@ from pydantic import BaseModel
 from fastapi import APIRouter, Depends, HTTPException, status
 import logging
 
+from open_webui.apps.retrieval.main import app
+from open_webui.apps.retrieval.process.process import process_file, ProcessFileForm
 from open_webui.apps.webui.models.knowledge import (
     Knowledges,
     KnowledgeUpdateForm,
@@ -12,7 +14,6 @@ from open_webui.apps.webui.models.knowledge import (
 )
 from open_webui.apps.webui.models.files import Files, FileModel
 from open_webui.apps.retrieval.vector.connector import VECTOR_DB_CLIENT
-from open_webui.apps.retrieval.main import process_file, ProcessFileForm
 
 
 from open_webui.constants import ERROR_MESSAGES
@@ -162,7 +163,7 @@ def add_file_to_knowledge_by_id(
 
     # Add content to the vector database
     try:
-        process_file(ProcessFileForm(file_id=form_data.file_id, collection_name=id))
+        process_file(app.state, ProcessFileForm(file_id=form_data.file_id, collection_name=id))
     except Exception as e:
         log.debug(e)
         raise HTTPException(
@@ -227,7 +228,7 @@ def update_file_from_knowledge_by_id(
 
     # Add content to the vector database
     try:
-        process_file(ProcessFileForm(file_id=form_data.file_id, collection_name=id))
+        process_file(app.state, ProcessFileForm(file_id=form_data.file_id, collection_name=id))
     except Exception as e:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
